cronwrap - add option parsing, add tweakable lockfile name
previously you could say which directory you wanted the lockfile in (to
avoid NFS hilarity, mostly) but not its actual name. however, you might
want to share a lock with another job, so specifying both parts is
useful.

ideally there would be just one option to specify both path and name but
backward compatibility is a thing even for a meagre shell wrapper.