cronwrap.lockf - use `lockf` for the locking
deliberately not done: autodetecting if `lockf` is available and using
it in place of `lockrun`. it's too much baggage for what is trying to
remain a simple script.

rather: the FreeBSD port will pick up the `cronwrap.lockf` variant and
install it as `cronwrap` and ... whatever happens elsewhere can cope.